requirements.txt
# Flask & Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Werkzeug==2.3.6
click==8.1.3

# Database
pymongo==4.4.1
mongoengine==0.27.0
python-dotenv==1.0.0

# Machine Learning & Data Processing
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
joblib==1.3.1
pickle5==0.0.12

# NLP & Text Processing
transformers==4.30.2
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.5.0
textblob==0.17.1

# Computer Vision & Image Processing
opencv-python==4.8.0
Pillow==10.0.0
imagehash==4.3.1
insightface==0.7.3
facenet-pytorch==2.5.0
scikit-image==0.21.0
imageio==2.32.0

# Web Scraping & API
instagrapi==2.0.0
requests==2.31.0
selenium==4.10.0
beautifulsoup4==4.12.2
lxml==4.9.2
httpx==0.24.1
tweepy==4.14.0

# Image Search & Hashing
faiss-cpu==1.7.4
faiss-gpu==1.7.4  # Optional: uncomment for GPU support
pyhash==0.9.3

# Report Generation
reportlab==4.0.4
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.1
python-dotenv==1.0.0

# Async & Concurrency
asyncio-contextmanager==1.0.0
aiohttp==3.8.5
celery==5.3.1
redis==5.0.0

# Monitoring & Logging
python-json-logger==2.0.7
loguru==0.7.0
sentry-sdk==1.28.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1
factory-boy==3.3.0
faker==19.2.0

# Code Quality & Linting
black==23.7.0
flake8==6.0.0
pylint==2.17.5
isort==5.12.0
mypy==1.4.1

# Utilities & Helpers
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
colorama==0.4.6
click==8.1.3
pydantic==2.0.0
marshmallow==3.19.0

# API Documentation
Werkzeug==2.3.6
flasgger==0.9.7.1

# Deployment & DevOps
gunicorn==21.2.0
gevent==23.7.0
python-multipart==0.0.6

# Optional: GPU Support
# tensorflow-gpu==2.13.0  # Uncomment for GPU
# torch-cuda  # CUDA toolkit

# Optional: Deepfake Detection
# deepfakes==0.0.1  # May need special setup
# mediapy==1.1.5

# Optional: Advanced Analytics
# plotly==5.16.1
# dash==2.12.1

# Development Tools
ipython==8.14.0
jupyter==1.0.0
notebook==7.0.0
ipdb==0.13.13